break down Landing page to components
